+++ src/main/java/src/Item/Item.java
@@ -1,4 +1,4 @@
-package src;
+package src.Item;
 
 public class Item {
     private String name;
