+++ src/main/java/test/ItemTest.java
@@ -3,8 +3,8 @@ package test;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import src.Item;
-import src.ItemDetail;
+import src.Item.Item;
+import src.Item.ItemDetail;
 
 public class ItemTest {
     private Item item;
